+++ packages/react-server/src/client.tsx
@@ -3,3 +3,4 @@
 export { Link, LinkForm } from "./lib/client/link";
 export { useRouter } from "./lib/client/router";
 export { routerRevalidate } from "./features/router/client";
+export { useActionState } from "./features/server-action/client";
