+++ packages/react-server/src/features/server-action/client.tsx
@@ -21,3 +21,19 @@ function ThrowActionError() {
   }
   return null;
 }
+
+// re-export React.useActionState since the type is not officially available yet
+export const useActionState: ReactUseActionState = (...args) =>
+  (React as any).useActionState(...args);
+
+// type is copied from ReactDOM.useFormState
+// https://github.com/facebook/react/pull/28491
+type ReactUseActionState = <State, Payload>(
+  action: (state: Awaited<State>, payload: Payload) => State | Promise<State>,
+  initialState: Awaited<State>,
+  permalink?: string,
+) => [
+  state: Awaited<State>,
+  dispatch: (payload: Payload) => void,
+  isPending: boolean,
+];
